Добавляет модуль валидации
