+++ index.html
@@ -237,10 +237,10 @@
       <button type="button" class="error__button">Попробовать снова</button>
     </div>
   </template>
-<script src="js/main.js" type="module"></script>
 <script src="vendor/pristine/pristine.min.js"></script>
 <script src="vendor/leaflet/leaflet.js"></script>
 <script src="vendor/leaflet/leaflet.js.map"></script>
 <script src="vendor/nouislider/nouislider.js"></script>
+<script src="js/main.js" type="module"></script>
 </body>
 </html>
