+++ js/main.js
@@ -2,5 +2,6 @@
 import './create-mocks.js';
 import './generation.js';
 import './slider.js';
+import './validation.js';
 import './map.js';
 import './forms.js';
